Add timestamp to cookies to force revalidation.
Previously, this was susceptible to replay attacks where the cookie is
still trusted even if the user has been de-authorized on the Google
side.  This means that the only time we'd actually re-auth with Google
would be if underpants were restarted.  This adds a timestamp to the
authenticated cookie payload and ensures that the timestamp is not too
old, which forces re-auth through Google once the cookie has matured.