Optionally add some security headers to responses.
This adds a new configurable in the JSON config file.  When enabled, a
few headers which can improve application security are added to all
proxied and self-generated HTTP responses.  The headers are:

    X-Frame-Options: SAMEORIGIN
    Cache-Control: private, no-cache
    Pragma: no-cache
    Strict-Transport-Security: max-age=16070400; includeSubDomains

The latter is only added if TLS certs are configured.

If not configured in the JSON, the old behaviour (not adding headers)
will be maintained.