Use new mux package for routing.
Go 1.9 introduced a breaking change in the way that http.ServeMux handles
host routes. As a result, underpants lost the ability to route on hosts
that included ports (i.e. protected:8080). This change removes the use
of http.ServeMux and uses a custom mux in the mux package to that underpants
works consistently reguardless of which version of Go it was compiled with.